feat(DateTime): combining components
