+++ src/components/DateTime/DateTime.tsx
@@ -18,8 +18,8 @@ import {
 export const cnDateTime = cn('DateTime');
 
 const viewMap: Record<DateTimePropView, DateTimeViewComponent> = {
-  oneMonth: DateTimeViewOneMonth,
-  twoMonths: DateTimeViewTwoMonths,
+  classic: DateTimeViewOneMonth,
+  book: DateTimeViewTwoMonths,
   slider: DateTimeViewSlider,
 };
 
@@ -46,12 +46,6 @@ export * from './DateTimeTime/DateTimeTime';
 export * from './DateTimeYearSlider/DateTimeYearSlider';
 export * from './DateTime10YearSlider/DateTime10YearSlider';
 export * from './DateTime100YearSlider/DateTime100YearSlider';
-export * from './DateTimeTypeYearViewClassic/DateTimeTypeYearViewClassic';
-export * from './DateTimeTypeYearViewBook/DateTimeTypeYearViewBook';
-export * from './DateTimeTypeYearViewSlider/DateTimeTypeYearViewSlider';
-export * from './DateTimeTypeMonthViewClassic/DateTimeTypeMonthViewClassic';
-export * from './DateTimeTypeMonthViewBook/DateTimeTypeMonthViewBook';
-export * from './DateTimeTypeMonthViewSlider/DateTimeTypeMonthViewSlider';
-export * from './DateTimeTypeDateViewClassic/DateTimeTypeDateViewClassic';
-export * from './DateTimeTypeDateViewBook/DateTimeTypeDateViewBook';
-export * from './DateTimeTypeDateViewSlider/DateTimeTypeDateViewSlider';
+export * from './DateTimeTypeYear/DateTimeTypeYear';
+export * from './DateTimeTypeMonth/DateTimeTypeMonth';
+export * from './DateTimeTypeDate/DateTimeTypeDate';
