--- src/components/DateTime/DateTimeTypeDateViewClassic/DateTimeTypeDateViewClassic.tsx
@@ -1,99 +1,0 @@
-import React from 'react';
-import { addMonths, startOfMonth } from 'date-fns';
-import ruLocale from 'date-fns/locale/ru';
-
-import { useFlag } from '../../../hooks/useFlag/useFlag';
-import { DateTimeMonth } from '../DateTimeMonth/DateTimeMonth';
-import { DateTimeToggler } from '../DateTimeToggler/DateTimeToggler';
-import { DateTimeTypeMonthViewClassic } from '../DateTimeTypeMonthViewClassic/DateTimeTypeMonthViewClassic';
-import {
-  DateTimeViewComponent,
-  getDaysOfMonth,
-  getDaysOfWeek,
-  getHandleSelectDate,
-  getMouthLabelWithYear,
-  isEqualDay,
-  useCurrentVisibleDate,
-} from '../helpers';
-import { cnDateTimeMixViewClassic } from '../mixs';
-
-export const DateTimeTypeDateViewClassic: DateTimeViewComponent = React.forwardRef((props, ref) => {
-  const {
-    className,
-    minDate,
-    maxDate,
-    value,
-    onChange,
-    onChangeRange,
-    currentVisibleDate: currentVisibleDateProp,
-    events,
-    locale = ruLocale,
-    ...otherProps
-  } = props;
-
-  const [changeMonth, { on, off }] = useFlag();
-
-  const [currentVisibleDate, setCurrentVisibleDate] = useCurrentVisibleDate({
-    currentVisibleDate: currentVisibleDateProp,
-    maxDate,
-    minDate,
-    value,
-    startOfUnit: startOfMonth,
-  });
-
-  const handleSelectDate = getHandleSelectDate({
-    minDate,
-    maxDate,
-    value,
-    onChange,
-    onChangeRange,
-    isEqualUnit: isEqualDay,
-  });
-
-  const daysOfMonth = getDaysOfMonth({
-    date: currentVisibleDate,
-    handleDayClick: handleSelectDate,
-    value,
-    events,
-    minDate,
-    maxDate,
-    locale,
-  });
-
-  const monthLabel = getMouthLabelWithYear(currentVisibleDate, locale);
-  const daysOfWeek = getDaysOfWeek(locale);
-  const nextMonthHandle = () => setCurrentVisibleDate(addMonths(currentVisibleDate, 1));
-  const prevMonthHandle = () => setCurrentVisibleDate(addMonths(currentVisibleDate, -1));
-
-  if (changeMonth) {
-    return (
-      <DateTimeTypeMonthViewClassic
-        {...otherProps}
-        ref={ref}
-        className={className}
-        minDate={minDate}
-        maxDate={maxDate}
-        locale={locale}
-        events={events}
-        currentVisibleDate={currentVisibleDate}
-        onChange={({ value }) => {
-          setCurrentVisibleDate(value);
-          off();
-        }}
-      />
-    );
-  }
-
-  return (
-    <div {...otherProps} ref={ref} className={cnDateTimeMixViewClassic(null, [className])}>
-      <DateTimeToggler
-        className={cnDateTimeMixViewClassic('Label')}
-        prevOnClick={prevMonthHandle}
-        nextOnClick={nextMonthHandle}
-        label={monthLabel}
-        onLabelClick={on}
-      />
-      <DateTimeMonth daysOfWeek={daysOfWeek} daysOfMonth={daysOfMonth} />
-    </div>
-  );
-});
