+++ src/components/DateTime/__stories__/examples/DateTimeExample/DateTimeExample.tsx
@@ -14,7 +14,7 @@ const events = [startOfWeek(new Date(), { locale: ruLocale }), new Date(), addDa
 export const DateTimeExampleView = () => {
   return (
     <StoryBookExample className={cnDocsDecorator('Section')}>
-      <DateTime view="oneMonth" />
+      <DateTime view="classic" />
     </StoryBookExample>
   );
 };
@@ -22,7 +22,7 @@ export const DateTimeExampleView = () => {
 export const DateTimeExampleViewTwo = () => {
   return (
     <StoryBookExample className={cnDocsDecorator('Section')}>
-      <DateTime view="twoMonths" />
+      <DateTime view="book" />
     </StoryBookExample>
   );
 };
