+++ src/components/DateTime/mixs/DateTimeMixLayout/DateTimeMixLayout.ts
@@ -1,0 +1,5 @@
+import './DateTimeMixLayout.css';
+
+import { cn } from '../../../../utils/bem';
+
+export const cnDateTimeMixLayout = cn('DateTimeMixLayout');
