--- src/components/DateTime/mixs/DateTimeMixViewBook/DateTimeMixViewBook.css
@@ -1,13 +1,0 @@
-.DateTimeMixViewBook {
-  display: flex;
-
-  &-Label {
-    margin-bottom: var(--space-l);
-  }
-
-  &-Page {
-    &:not(:last-child) {
-      margin-right: var(--space-m);
-    }
-  }
-}
