--- src/components/DateTime/mixs/DateTimeMixViewBook/DateTimeMixViewBook.ts
@@ -1,5 +1,0 @@
-import './DateTimeMixViewBook.css';
-
-import { cn } from '../../../../utils/bem';
-
-export const cnDateTimeMixViewBook = cn('DateTimeMixViewBook');
