--- src/components/DateTime/mixs/DateTimeMixViewClassic/DateTimeMixViewClassic.css
@@ -1,8 +1,0 @@
-.DateTimeMixViewClassic {
-  --date-time-cell-width: calc(var(--control-height-s) + (var(--space-3xs) * 2));
-  width: calc(var(--date-time-cell-width) * 7);
-
-  &-Label {
-    margin-bottom: var(--space-l);
-  }
-}
