--- src/components/DateTime/mixs/DateTimeMixViewClassic/DateTimeMixViewClassic.ts
@@ -1,5 +1,0 @@
-import './DateTimeMixViewClassic.css';
-
-import { cn } from '../../../../utils/bem';
-
-export const cnDateTimeMixViewClassic = cn('DateTimeMixViewClassic');
