--- src/components/DateTime/mixs/DateTimeMixViewSlider/DateTimeMixViewSlider.css
@@ -1,20 +1,0 @@
-.DateTimeMixViewSlider {
-  &-Slider {
-    margin-bottom: var(--space-l);
-  }
-
-  &-Label {
-    display: block;
-    margin-bottom: var(--space-s);
-  }
-
-  &-PageWrapper {
-    display: flex;
-  }
-
-  &-Page {
-    &:not(:last-child) {
-      margin-right: var(--space-m);
-    }
-  }
-}
