--- src/components/DateTime/mixs/DateTimeMixViewSlider/DateTimeMixViewSlider.ts
@@ -1,5 +1,0 @@
-import './DateTimeMixViewSlider.css';
-
-import { cn } from '../../../../utils/bem';
-
-export const cnDateTimeMixViewSlider = cn('DateTimeMixViewSlider');
