+++ src/components/DateTime/mixs/index.ts
@@ -1,3 +1,1 @@
-export * from './DateTimeMixViewClassic/DateTimeMixViewClassic';
-export * from './DateTimeMixViewBook/DateTimeMixViewBook';
-export * from './DateTimeMixViewSlider/DateTimeMixViewSlider';
+export * from './DateTimeMixLayout/DateTimeMixLayout';
